+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/sqlCompiler/circuit/operator/InputColumnMetadata.java
@@ -9,30 +9,26 @@ import javax.annotation.Nullable;
  * Metadata describing an input table column.
  */
 public class InputColumnMetadata {
-    /**
-     * Column name.
-     */
+    /** Column name. */
     public final String name;
-    /**
-     * Column type.
-     */
+    /** Column type. */
     public final DBSPType type;
-    /**
-     * True if the column is a primary key.
-     */
+    /** True if the column is part of a primary key. */
     public final boolean isPrimaryKey;
-    /**
-     * Lateness, if declared.  Should be a constant expression.
-     */
+    /** Lateness, if declared.  Should be a constant expression. */
     @Nullable
     public final DBSPExpression lateness;
+    /** Default value, if declared.  Should be a constant expression */
+    @Nullable
+    public final DBSPExpression defaultValue;
 
     public InputColumnMetadata(String name, DBSPType type, boolean isPrimaryKey,
-                               @Nullable DBSPExpression lateness) {
+                               @Nullable DBSPExpression lateness, @Nullable DBSPExpression defaultValue) {
         this.name = name;
         this.type = type;
         this.isPrimaryKey = isPrimaryKey;
         this.lateness = lateness;
+        this.defaultValue = defaultValue;
     }
 
     public String getName() {