+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - WebConsole: Support big numeric SQL types in Data Browser, Data Import and Data Generator (#851)
 - WebConsole: Display SQL types in Data inspect and insert tables
+- SQL: parser support for 'DEFAULT' column values in DDL (#1061)
 
 ## [0.4.0] - 2023-11-21